@@ -11,46 +11,68 @@
     "electron:build": "tsc && vite build && electron-builder"
   },
   "dependencies": {
+    "5": "^0.0.1",
+    "6": "^0.0.1",
     "@fortawesome/fontawesome-svg-core": "^6.7.1",
     "@fortawesome/free-solid-svg-icons": "^6.7.1",
     "@fortawesome/react-fontawesome": "^0.2.2",
+    "@jimp/core": "^1.6.0",
+    "@jimp/custom": "^0.22.12",
     "@tensorflow-models/coco-ssd": "^2.2.3",
     "@tensorflow/tfjs": "^4.22.0",
     "@types/express": "^5.0.0",
     "@types/mongoose": "^5.11.97",
     "@types/node": "^22.10.1",
     "@types/react-router-dom": "^5.3.3",
-    "axios": "^1.7.9",
+    "axios": "^1.8.4",
+    "chroma-js": "^3.1.2",
+    "color": "^5.0.0",
     "color-diff": "^1.4.0",
     "color-thief-node": "^1.0.4",
+    "colornames": "^1.1.1",
     "colorthief": "^2.6.0",
     "cors": "^2.8.5",
+    "dotenv": "^16.4.7",
     "express": "^4.21.1",
     "fabric": "^4.6.0",
+    "get-image-colors": "^4.0.1",
+    "image-js": "^0.36.0",
     "imagetracerjs": "github:jankovicsandras/imagetracerjs",
-    "jspdf": "^2.5.2",
+    "jimp": "^1.6.0",
+    "jpeg-js": "^0.4.4",
+    "jspdf": "^3.0.1",
+    "minimist": "^1.2.8",
     "ml-kmeans": "^6.0.0",
-    "mongoose": "^8.8.3",
+    "mongoose": "^8.12.2",
+    "node-vibrant": "^4.0.3",
+    "npm-force-resolutions": "^0.0.10",
     "opencv.js": "^1.2.1",
+    "phin": "^3.7.1",
     "pixi.js": "^8.6.6",
     "potrace": "^2.1.8",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-draggable": "^4.4.6",
     "react-resizable": "^3.0.5",
     "react-router-dom": "^7.0.1",
-    "ts-node": "^10.9.2"
+    "responselike": "^3.0.0",
+    "sharp": "^0.33.5",
+    "tinycolor2": "^1.6.0",
+    "tough-cookie": "^5.1.2",
+    "vite-plugin-commonjs": "^0.10.4"
   },
   "devDependencies": {
     "@types/color-diff": "^1.2.5",
-    "@types/fabric": "^5.3.9",
-    "@types/react": "^18.0.27",
-    "@types/react-dom": "^18.0.10",
+    "@types/fabric": "^5.3.10",
+    "@types/react": "^18.3.19",
+    "@types/react-dom": "^18.3.5",
     "@types/react-resizable": "^3.0.8",
-    "@vitejs/plugin-react": "^3.1.0",
-    "electron": "^34.0.2",
+    "@types/responselike": "^1.0.3",
+    "@vitejs/plugin-react": "^4.3.4",
+    "electron": "^34.3.4",
     "electron-builder": "^25.1.8",
-    "typescript": "^4.9.5",
-    "vite": "^4.1.0"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.2",
+    "vite": "^6.2.2"
   }
 }
